requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langgraph>=0.2.0
chromadb>=0.5.0
openai>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.31.0
pytest>=8.0.0
httpx>=0.27.0